feat(web): normalize email
