+++ apps/web/package.json
@@ -34,6 +34,7 @@
     "@babel/plugin-proposal-optional-chaining": "^7.20.7",
     "@babel/plugin-transform-react-display-name": "^7.18.6",
     "@babel/plugin-transform-runtime": "^7.23.2",
+    "@clerk/clerk-js": "^5.10.0",
     "@clerk/clerk-react": "^5.2.5",
     "@clerk/themes": "^2.1.10",
     "@editorjs/editorjs": "^2.19.3",
